+++ client/src/components/game/mechanics/ghosts/collisions/pacmanDeath/checkPacmanLives.js
@@ -1,5 +1,5 @@
-import endGame from "../endGame";
-import resetAfterDeath from "../resetAfterDeath";
+import endGame from "./endGame";
+import resetAfterDeath from "./resetAfterDeath";
 
 export default function checkPacmanLives(
   pacman,
