+++ client/src/components/game/mechanics/ghosts/collisions/pacmanDeath/checkPacmanLives.test.js
