+++ client/src/components/game/mechanics/ghosts/collisions/pacmanDeath/drawBoard.js
