+++ client/src/components/game/mechanics/ghosts/collisions/pacmanDeath/resetAfterDeath.js
@@ -1,4 +1,4 @@
-import playGame from "../../playGame";
+import playGame from "../../../playGame";
 
 export default function resetAfterDeath(
   pacman,
