+++ client/src/components/game/mechanics/ghosts/collisions/pacmanDeath/resetAfterGameOver.js
