+++ client/src/components/game/mechanics/ghosts/collisions/pacmanDeath/runDeathAnimation.js
