+++ client/src/components/game/mechanics/ghosts/collisions/pacmanDeath/saveScore.js
