+++ client/src/components/game/mechanics/pellets/levelUpAnimation/resetAfterLevelUp.test.js
