+++ client/src/components/game/mechanics/pellets/levelUpAnimation/runLevelUpAnimation.js
